TCMK Faz 4: x86_64 portu (Long Mode, 4 seviyeli sayfalama, ELF64, syscall)
Cekirdek artik iki mimaride calisiyor: make ARCH=i386 / ARCH=x86_64.

Mimari soyutlama:
- Ortak katmanlar donanima her zaman `arch::cpu` uzerinden erisir; somut
  `arch::i386` adi ortak koddan tamamen kalkti
- SyscallFrame sozlesmesi `usize`e tasindi -- POSIX/NT cevirmenleri iki
  mimaride de tek kod tabaniyla calisiyor
- gdt/ ve idt/ mimariye ozel alt modullere bolundu

x86_64 tarafi:
- Multiboot2 + 32-bit stub'dan Long Mode'a gecis (PML4/PDPT/PD, PAE,
  EFER.LME, CR0.PG, GDT64 + uzak atlama)
- 4 seviyeli sayfalama; kullanici bolgesi icin 2 MiB'lik girdi 4 KiB'lik
  tabloya bolunerek sayfa sayfa aciliyor
- 64-bit GDT + TSS (rsp0), 16 baytlik IDT girdileri
- `syscall`/`sysret` altyapisi (EFER.SCE, STAR/LSTAR/SFMASK)
- ELF64 yukleyici + x86_64 userland ureteci

Yol boyunca cikan uc gercek hata (README "Bilinen Tuzaklar"a eklendi):

1. Linux syscall numaralari mimariye gore degisir (i386 write=4/exit=1,
   x86_64 write=1/exit=60). Tek kumeyle gecistirilince x86_64 userland
   `write`(=1) cagirdi, cekirdek bunu i386'nin `exit`'i sandi ve program
   hicbir sey yazmadan "cikis kodu 1" ile sonlandi.
2. `syscall` komutu Ring 0'dan cagrilamaz: donusu `sysretq` ve o HER ZAMAN
   Ring 3'e doner. Cekirdek ici syscall3 yardimcisi x86_64'te de int 0x80
   kullaniyor.
3. Boot stub'inda 2 MiB identity girdilerine User biti konmasi cekirdegin
   tamamini Ring 3'e aciyordu (izolasyon testi kernel@0x100000=true
   veriyordu). Bit kaldirildi; acma isi protect_user_range'e birakildi.

Dogrulama -- her iki mimaride de:
  Ring 3 CPL=3 (x86_64: CS=0x23 SS=0x1b, gdb ile dogrulandi)
  izolasyon: user=true kernel=false heap=false
  guvenlik: sys_open(cekirdek isaretcisi) -> -14 EFAULT
  i386 ayrica ELF32 + PE32 cift uyumlulugunu korumaya devam ediyor

+++ kernel/src/arch/i386/regs.rs
@@ -42,12 +42,21 @@ impl SyscallFrame {
     }
 
     /// i386 Linux ABI: EBX, ECX, EDX, ESI, EDI = arg1..arg5.
-    pub fn args(&self) -> [u32; 5] {
-        [self.ebx, self.ecx, self.edx, self.esi, self.edi]
+    ///
+    /// Donus tipi `usize`: ortak katmanlar (POSIX/NT cevirmenleri) boylece
+    /// i386 ve x86_64'te ayni kodla calisir.
+    pub fn args(&self) -> [usize; 5] {
+        [
+            self.ebx as usize,
+            self.ecx as usize,
+            self.edx as usize,
+            self.esi as usize,
+            self.edi as usize,
+        ]
     }
 
     /// Donus degeri EAX uzerinden verilir.
-    pub fn set_return(&mut self, value: u32) {
-        self.eax = value;
+    pub fn set_return(&mut self, value: usize) {
+        self.eax = value as u32;
     }
 }
